% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{add_field}
\alias{add_field}
\title{Add a field to a table in a \link{RelDataModel}}
\usage{
add_field(x, tableName, name, type, nullable, unique, comment)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{tableName}{the name of the table to modify (a single character)}

\item{name}{the name of the field to add (a single character)}

\item{type}{the type of the field (a single character)}

\item{nullable}{if the field is nullable (a single logical)}

\item{unique}{if the values are unique (a single logical)}

\item{comment}{a description (a single character)}
}
\value{
A \link{RelDataModel}
}
\description{
Add a field to a table in a \link{RelDataModel}
}
