% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{add_index}
\alias{add_index}
\title{Add an index to a table in a \link{RelDataModel}}
\usage{
add_index(x, tableName, fieldNames, unique)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{tableName}{the name of the table to modify (a single character)}

\item{fieldNames}{the names of the fields to include in the index}

\item{unique}{a logical indicating if the indexed values are unique}
}
\value{
A \link{RelDataModel}
}
\description{
Add an index to a table in a \link{RelDataModel}
}
