% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{add_table}
\alias{add_table}
\title{Add a table to a \link{RelDataModel}}
\usage{
add_table(x, newTable)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{newTable}{the name of the new table or a \link{RelTableModel}}
}
\value{
A \link{RelDataModel}
}
\description{
Add a table to a \link{RelDataModel}
}
