% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{addTable.RelDataModel}
\alias{addTable.RelDataModel}
\title{Add a table to a \link{RelDataModel}}
\usage{
\method{addTable}{RelDataModel}(x, newTable)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{newTable}{the name of the new table or a \link{RelTableModel}}
}
\value{
A \link{RelDataModel}
}
\description{
Add a table to a \link{RelDataModel}
}
