% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TYPEDEF.R
\name{as_type}
\alias{as_type}
\title{Convert an object into a specific type}
\usage{
as_type(x, type)
}
\arguments{
\item{x}{an object to convert}

\item{type}{the targeted type}
}
\description{
Convert an object into a specific type
}
