% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{auto_layout}
\alias{auto_layout}
\title{Pre-compute \link{RelDataModel} layout when missing any x or y table position}
\usage{
auto_layout(
  x,
  layout = "layout_nicely",
  lengthMultiplier = 40 * length(x),
  force = FALSE
)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{layout}{character name of igraph layout function to use
(Default: "layout_nicely").}

\item{lengthMultiplier}{a numeric value to scale x and y coordinate
(default: 40*length(x))}

\item{force}{if TRUE autolayout even if all tables have coordinates
(default: FALSE)}
}
\value{
A \link{RelDataModel}
}
\description{
Pre-compute \link{RelDataModel} layout when missing any x or y table position
}
