% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_relational_data.R
\name{bring_back_RelDataModel}
\alias{bring_back_RelDataModel}
\title{Bring back an autosaved \link{RelDataModel}}
\usage{
bring_back_RelDataModel(name)
}
\arguments{
\item{name}{The name of the autosaved \link{RelDataModel} to bring back.
Available autosaved \link{RelDataModel} can be listed using
the \code{\link[=list_autosaved_RelDataModel]{list_autosaved_RelDataModel()}}}
}
\description{
Bring back an autosaved \link{RelDataModel}
}
