% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{c.RelDataModel}
\alias{c.RelDataModel}
\title{Merge \link{RelDataModel} objects}
\usage{
\method{c}{RelDataModel}(..., checkFK = TRUE)
}
\arguments{
\item{...}{\link{RelDataModel} objects}

\item{checkFK}{a logical indicating if foreign keys should be checked
(default: TRUE)}
}
\value{
A \link{RelDataModel} objects
}
\description{
Merge \link{RelDataModel} objects
}
