% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{check_foreign_keys}
\alias{check_foreign_keys}
\title{Check that all defined foreign keys are available}
\usage{
check_foreign_keys(x, ...)
}
\description{
Check that all defined foreign keys are available
}
