% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TYPEDEF.R
\name{check_types}
\alias{check_types}
\title{Check if a set of types is supported}
\usage{
check_types(x)
}
\arguments{
\item{x}{a character vector of types to be checked}
}
\description{
Check if a set of types is supported
}
