% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{checkForeignKeys.RelDataModel}
\alias{checkForeignKeys.RelDataModel}
\title{Check the availability of foreign keys}
\usage{
\method{checkForeignKeys}{RelDataModel}(x)
}
\value{
Nothing. The function throws an error if there is an issue with
foreign keys.
}
\description{
Check the availability of foreign keys
}
