% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TYPEDEF.R
\name{checkTypes}
\alias{checkTypes}
\title{Check if a set of types is supported}
\usage{
checkTypes(x)
}
\arguments{
\item{x}{a character vector of types to be checked}
}
\description{
Check if a set of types is supported
}
