% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_relational_data.R
\name{clean_autosaved_RelDataModels}
\alias{clean_autosaved_RelDataModels}
\title{Remove all autosaved \link{RelDataModel}}
\usage{
clean_autosaved_RelDataModels()
}
\description{
Remove all autosaved \link{RelDataModel}
}
