% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_relational_data.R
\name{clean_RelDataModel_in_cache}
\alias{clean_RelDataModel_in_cache}
\title{Remove all cached \link{RelDataModel} from .GlobalEnv}
\usage{
clean_RelDataModel_in_cache()
}
\description{
Remove all cached \link{RelDataModel} from .GlobalEnv
}
