% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelTableModel.R
\name{col_types}
\alias{col_types}
\title{Get the types of the columns of a \link{RelTableModel} object}
\usage{
col_types(x)
}
\arguments{
\item{x}{a \link{RelTableModel} object}
}
\value{
A col_spec object with the type of each column
}
\description{
Get the types of the columns of a \link{RelTableModel} object
}
