% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{confront_data.RelDataModel}
\alias{confront_data.RelDataModel}
\title{Confront a \link{RelDataModel} to actual data}
\usage{
\method{confront_data}{RelDataModel}(
  x,
  data = list(),
  paths = NULL,
  returnData = FALSE,
  n_max = Inf,
  checks = if (n_max == Inf) {     c("unique", "not nullable", "foreign keys") } else { 
       as.character() },
  delim = "\\t",
  ...
)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{paths}{a character vector with file paths.
The file \link{basename} without extension
will be considered as the table name.}

\item{returnData}{a logical indicating if the data should be returned
with the report (default: FALSE).}

\item{n_max}{maximum number of records to read (default: Inf).}

\item{checks}{a character vector with the name of optional checks to be done
(Default: if n_max==Inf ==> all of
them c("unique", "not nullable", "foreign keys"),
else ==> none)}

\item{delim}{single character used to separate fields within a record
(default: "\\t")}

\item{...}{supplementary parameters for the \link{read_delim} function.}
}
\value{
A report as a list
}
\description{
Confront a \link{RelDataModel} to actual data
}
