% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelTableModel.R
\name{confront_data.RelTableModel}
\alias{confront_data.RelTableModel}
\title{Confront a \link{RelTableModel} to actual data}
\usage{
\method{confront_data}{RelTableModel}(x, d, checks = c("unique", "not nullable"))
}
\arguments{
\item{x}{a \link{RelTableModel}}

\item{d}{a data frame}

\item{checks}{a character vector with the name of optional checks to be done
(Default: all of them c("unique", "not nullable"))}
}
\value{
A report as a list
}
\description{
Confront a \link{RelTableModel} to actual data
}
