% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TYPEDEF.R
\name{conv_type_ref}
\alias{conv_type_ref}
\title{Convert a set of types from or to R supported types}
\usage{
conv_type_ref(x, from = NULL, to = NULL, ignore.case = TRUE)
}
\arguments{
\item{x}{a character vector of types to be converted. If from is not null,
x should be a set of valid types in the from reference. If to is not null,
x should be a set of supported R types (SUPPTYPES).}

\item{from}{a character vector of length one: the type reference
(\link{list_type_ref}) of x}

\item{to}{a character vector of length one: the targeted type reference
(\link{list_type_ref})}

\item{ignore.case}{should case be ignored when converting `from``
type reference (default: TRUE)}
}
\description{
Convert a set of types from or to R supported types
}
\details{
Only \code{from} XOR \code{to} should be set
}
