% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{copy_fields}
\alias{copy_fields}
\title{Copy fields from one table to another in a \link{RelDataModel}}
\usage{
copy_fields(x, from, to, fields)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{from}{the name of the table from which the fields are taken}

\item{to}{the name of the table to which the fields are copied}

\item{fields}{the names of the fields to copy}
}
\value{
A \link{RelDataModel}
}
\description{
Copy fields from one table to another in a \link{RelDataModel}
}
