% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelTableModel.R
\name{correct_constraints}
\alias{correct_constraints}
\title{Correct the constraints of a table to make them consistent}
\usage{
correct_constraints(x, createPKIndex = FALSE)
}
\arguments{
\item{x}{a \link{RelTableModel} object}

\item{createPKIndex}{should an index be create for the primary key
(default: FALSE)}
}
\description{
Correct the constraints of a table to make them consistent
}
