% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelTableModel.R
\name{correct_constraints.RelTableModel}
\alias{correct_constraints.RelTableModel}
\title{Correct the constraints of a table to make them consistent}
\usage{
\method{correct_constraints}{RelTableModel}(x)
}
\arguments{
\item{x}{a \link{RelTableModel} object}
}
\description{
Correct the constraints of a table to make them consistent
}
