% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_confrontation_report.R
\name{format_confrontation_report}
\alias{format_confrontation_report}
\title{Format confrontation report in markdown format}
\usage{
format_confrontation_report(
  cr,
  title = "Model",
  bgSuccess = "green",
  txSuccess = "black",
  bgFailure = "red",
  txFailure = "white",
  bgMessage = "#FFBB33",
  txMessage = "white"
)
}
\arguments{
\item{cr}{the confrontation report from \link{confront_data.RelDataModel}}
}
\description{
Format confrontation report in markdown format
}
