% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_foreign_keys}
\alias{get_foreign_keys}
\title{Get a foreign key table from an object}
\usage{
get_foreign_keys(x)
}
\arguments{
\item{x}{a \link{RelTableModel} or a \link{RelDataModel}}
}
\value{
A tibble with the following fields:
\itemize{
\item from: the origin of the key
\item ff: the key fields in from
\item to: the target of the key
\item tf: the key fields in to
\item fmin: minimum cardinality of from
\item fmax: maximum cardinality of from
\item tmin: minimum cardinality of to
\item tmax: maximum cardinality of to
}
}
\description{
Get a foreign key table from an object
}
