% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelTableModel.R
\name{get_foreign_keys.RelTableModel}
\alias{get_foreign_keys.RelTableModel}
\title{Get foreign keys from \link{RelTableModel}}
\usage{
\method{get_foreign_keys}{RelTableModel}(x)
}
\arguments{
\item{x}{a \link{RelTableModel}}
}
\value{
A tibble with the following fields:
\itemize{
\item from: the origin of the key
\item ff: the key fields in from
\item to: the target of the key
\item tf: the key fields in to
\item fmin: minimum cardinality of from
\item fmax: maximum cardinality of from
\item tmin: minimum cardinality of to
\item tmax: maximum cardinality of to
}
}
\description{
Get foreign keys from \link{RelTableModel}
}
