% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{identical_RelDataModel}
\alias{identical_RelDataModel}
\title{Check if two \link{RelDataModel} are identical}
\usage{
identical_RelDataModel(x, y, ...)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{y}{a \link{RelDataModel}}

\item{...}{additional parameters for \code{\link[=identical_RelTableModel]{identical_RelTableModel()}}}
}
\value{
A logical: TRUE if the 2 models are identical
}
\description{
Check if two \link{RelDataModel} are identical
}
