% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelTableModel.R
\name{identical_RelTableModel}
\alias{identical_RelTableModel}
\title{Check if two \link{RelTableModel} are identical}
\usage{
identical_RelTableModel(x, y, includeDisplay = TRUE)
}
\arguments{
\item{x}{a \link{RelTableModel}}

\item{y}{a \link{RelTableModel}}

\item{includeDisplay}{a single logical (default: TRUE) indicating if
the display should be included in the comparison}
}
\value{
A logical: TRUE if the 2 models are identical
}
\description{
Check if two \link{RelTableModel} are identical
}
