% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelTableModel.R
\name{indexTable.RelTableModel}
\alias{indexTable.RelTableModel}
\title{List indexes of a \link{RelTableModel} object}
\usage{
\method{indexTable}{RelTableModel}(x)
}
\arguments{
\item{x}{a \link{RelTableModel} object}
}
\value{
A tibble with the following columns:
\itemize{
\item \strong{index}: an integer corresponding to the index number
\item \strong{field}: a character corresponding to field belonging to the index
\item \strong{unique}: a logical indicating the uniqueness of the field
}
}
\description{
List indexes of a \link{RelTableModel} object
}
