% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelTableModel.R
\name{is.MatrixModel}
\alias{is.MatrixModel}
\title{Check if the object is a \link{RelTableModel} matrix object}
\usage{
is.MatrixModel(x)
}
\arguments{
\item{x}{any object}
}
\value{
A single logical: TRUE if x is a \link{RelTableModel} matrix object
}
\description{
A matrix model is a special \link{RelTableModel} object with 3 and only 3 fields:
2 of types 'row' and 'column' and the 3rd of your choice.
}
