% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{is.RelDataModel}
\alias{is.RelDataModel}
\title{Check if the object is  a \link{RelDataModel} object}
\usage{
is.RelDataModel(x)
}
\arguments{
\item{x}{any object}
}
\value{
A single logical: TRUE if x is a \link{RelDataModel} object
}
\description{
Check if the object is  a \link{RelDataModel} object
}
