% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelTableModel.R
\name{is.RelTableModel}
\alias{is.RelTableModel}
\title{Check the object is  a \link{RelTableModel} object}
\usage{
is.RelTableModel(x)
}
\arguments{
\item{x}{any object}
}
\value{
A single logical: TRUE if x is a \link{RelTableModel} object
}
\description{
Check the object is  a \link{RelTableModel} object
}
