% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_MM.R
\name{is_MM}
\alias{is_MM}
\title{Identify if a file is in MatrixMarket text format}
\usage{
is_MM(file)
}
\arguments{
\item{file}{the file to read}
}
\value{
A logical. If FALSE, the first line of the file is returned as
an attribute named "r1": \code{attr(is_MM, "r1")}
}
\description{
Identify if a file is in MatrixMarket text format
}
