% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{length.RelDataModel}
\alias{length.RelDataModel}
\title{Get the number of tables in a \link{RelDataModel} object}
\usage{
\method{length}{RelDataModel}(x)
}
\description{
Get the number of tables in a \link{RelDataModel} object
}
