% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{lengths}
\alias{lengths}
\title{Lengths of object elements}
\usage{
lengths(x, use.names = TRUE)
}
\arguments{
\item{x}{an object. If there is no method implemented for this object,
the \code{\link[base:lengths]{base::lengths()}} function is used.}

\item{use.names}{logical indicating if the result should
inherit the names from x.}
}
\value{
A non-negative integer of length length(x), except when any element
has a length of more than 2^31 - 1 elements, when it returns a double vector.
When use.names is true, the names are taken from the names on x, if any.
}
\description{
Lengths of object elements
}
\seealso{
\code{\link[base:lengths]{base::lengths()}}
}
