% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_relational_data.R
\name{list_autosaved_RelDataModel}
\alias{list_autosaved_RelDataModel}
\title{List autosaved \link{RelDataModel}}
\usage{
list_autosaved_RelDataModel()
}
\description{
List autosaved \link{RelDataModel}
}
\seealso{
\code{\link[=clean_autosaved_RelDataModels]{clean_autosaved_RelDataModels()}} to clean this list.
}
