% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RelDataModel.R
\name{modelToVn}
\alias{modelToVn}
\title{VisNetwork representation of a \link{RelDataModel} object}
\usage{
modelToVn(
  model,
  border = "black",
  color = "lightgrey",
  highlightBorder = "orange"
)
}
\description{
Internal function
}
