% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TYPEDEF.R
\name{norm_type_ref}
\alias{norm_type_ref}
\title{Normalize type names}
\usage{
norm_type_ref(x, typeRef, ignore.case = TRUE)
}
\arguments{
\item{x}{a character vector to normalize}

\item{typeRef}{a character vector of length one: the type
reference (\link{list_type_ref})}

\item{ignore.case}{should case be ignored (default: TRUE)}
}
\description{
Normalize type names
}
