% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TYPEDEF.R
\name{normTypeRef}
\alias{normTypeRef}
\title{Normalize type names}
\usage{
normTypeRef(x, typeRef)
}
\arguments{
\item{x}{a character vector to normalize}

\item{typeRef}{a character vector of length one: the type
reference (\code{\link{listTypeRef}})}
}
\description{
Normalize type names
}
