% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{order_indexes}
\alias{order_indexes}
\title{Order index in a table in a \link{RelDataModel}}
\usage{
order_indexes(x, tableName, order)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{tableName}{the name of the table to modify (a single character)}

\item{order}{a vector of integers all in (1:number_of_indexes)}
}
\value{
A \link{RelDataModel}
}
\description{
Order index in a table in a \link{RelDataModel}
}
