% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RelDataModel.R
\name{plot.RelDataModel}
\alias{plot.RelDataModel}
\title{Plot a \link{RelDataModel} object}
\usage{
\method{plot}{RelDataModel}(x, ...)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{color}{default table background color}

\item{border}{border color (single character)}

\item{highlightBorder}{color of highlighted borders}
}
\description{
This function draw a visNetwork of the \link{RelDataModel}.
}
