% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_MM.R
\name{read_named_MM_header}
\alias{read_named_MM_header}
\title{Read the header of a named sparse matrix in MatrixMarket text format}
\usage{
read_named_MM_header(file, guess_max = 20)
}
\arguments{
\item{file}{the file to read}

\item{guess_max}{the number of lines to read to find the header.
(4 should be sufficient. Default: 20)}
}
\value{
A list with the following fields:
\itemize{
\item rownames: a character vector with the matrix row names
\item colnames: a character vector with the matrix column names
\item rows: the number of matrix rows
\item columns: the number of matrix columns
\item values: the number of values in the matrix
\item header_length: the number of lines in the header
}
}
\description{
Read the header of a named sparse matrix in MatrixMarket text format
}
