% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSQLDataModel.R
\name{read_SQL_data_model}
\alias{read_SQL_data_model}
\alias{readSQLDataModel}
\title{Read a data model from an SQL file from the MySQL Workbench}
\usage{
read_SQL_data_model(f, typeRef = "MySQLWB")

readSQLDataModel(...)
}
\arguments{
\item{f}{the SQL file to read}

\item{typeRef}{the reference for type conversion
(Default: "MySQLWB"; see \code{\link[=listTypeRef]{listTypeRef()}})}

\item{...}{params for \code{read_SQL_data_model}}
}
\value{
A \link{RelDataModel} object
}
\description{
Read a data model from an SQL file from the MySQL Workbench
}
\section{Functions}{
\itemize{
\item \code{readSQLDataModel}: Deprecated version of read_SQL_data_model
}}

