% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSQLDataModel.R
\name{readSQLDataModel}
\alias{readSQLDataModel}
\title{Read a data model from an SQL file from the MySQL Workbench}
\usage{
readSQLDataModel(f, typeRef = "MySQLWB")
}
\arguments{
\item{f}{the SQL file to read}

\item{typeRef}{the reference for type conversion
(Default: "MySQLWB"; see \code{\link[=listTypeRef]{listTypeRef()}})}
}
\value{
A \link{RelDataModel} object
}
\description{
Read a data model from an SQL file from the MySQL Workbench
}
