% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_relational_data.R
\name{recover_RelDataModel}
\alias{recover_RelDataModel}
\title{Recover an autosaved \link{RelDataModel}}
\usage{
recover_RelDataModel(name = NA)
}
\arguments{
\item{name}{The name of the autosaved \link{RelDataModel} to bring back.
Available autosaved \link{RelDataModel} can be listed using
the \code{\link[=list_autosaved_RelDataModel]{list_autosaved_RelDataModel()}}. If NA (default) the latest model is
returned.}
}
\description{
Recover an autosaved \link{RelDataModel}
}
