% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{RelDataModel}
\alias{RelDataModel}
\title{Create a RelDataModel object}
\usage{
RelDataModel(l, checkFK = TRUE)
}
\arguments{
\item{l}{the list of table models (\code{\link{RelTableModel}} objects)}

\item{checkFK}{a logical indicating if foreign keys should be checked}
}
\value{
A RelDataModel object.
}
\description{
Create a RelDataModel object
}
