% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelTableModel.R
\name{RelTableModel}
\alias{RelTableModel}
\title{Create a RelTableModel object}
\usage{
RelTableModel(l)
}
\arguments{
\item{l}{the list with the following elements:\itemize{
\item{tableName: a character vector of length one}
\item{fields: a tibble with the follwoing columns:\itemize{
   \item{name: character}
   \item{type: character}
   \item{nullable: logical}
   \item{comment:  character}
}}
\item{primaryKey: a character vector of any length. All
values should be in fields$name}
\item{foreignKeys: a list of foreign keys. Each foreigned key is defined
as a list with the following elements: \itemize{
   \item{refTable: a character vector of lenght one (the referenced table)}
   \item{key: a tibble with a "from" and a "to" columns}
}}
\item{indexes: a list of indexes. Each index is defined by
3 columns:\itemize{
    \item{field: character (all in fields$name)}
    \item{order: character}
    \item{unique: logical}
}}
}}
}
\value{
A RelTableModel object.
}
\description{
Create a RelTableModel object
}
