% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{remove_field}
\alias{remove_field}
\title{Remove a field from a table in a \link{RelDataModel}}
\usage{
remove_field(x, tableName, fieldName, rmForeignKeys = FALSE)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{tableName}{the name of the table to modify (a single character)}

\item{fieldName}{the name of the field to remove (a single character)}

\item{rmForeignKeys}{a single logical indicating if the corresponding foreign
keys should be removed. If FALSE (default), the function will throw an
error if it encounter a foreign key using the field.}
}
\value{
A \link{RelDataModel}
}
\description{
Remove a field from a table in a \link{RelDataModel}
}
