% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{remove_foreign_key}
\alias{remove_foreign_key}
\title{Remove a foreign key between two tables}
\usage{
remove_foreign_key(x, fromTable, fromFields, toTable, toFields)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{fromTable}{the name of the referencing table}

\item{fromFields}{the name of the referencing fields}

\item{toTable}{the name of the referenced table}

\item{toFields}{the names of the referenced fields}
}
\value{
A \link{RelDataModel}
}
\description{
Remove a foreign key between two tables
}
