% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{remove_index.RelDataModel}
\alias{remove_index.RelDataModel}
\title{Remove an index from a table in a \link{RelDataModel}}
\usage{
\method{remove_index}{RelDataModel}(x, tableName, fieldNames)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{tableName}{the name of the table to modify (a single character)}

\item{fieldNames}{the names of the fields composing the index}
}
\value{
A \link{RelDataModel}
}
\description{
Remove an index from a table in a \link{RelDataModel}
}
