% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{remove_table}
\alias{remove_table}
\title{Remove a table from a \link{RelDataModel}}
\usage{
remove_table(x, tableName, rmForeignKeys = FALSE)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{tableName}{the name of the table to remove}

\item{rmForeignKeys}{if TRUE, remove foreign keys which are not
available after extraction. If FALSE (default) the function will throw an
error if any foreign keys does not exist in the extracted RelDataModel.}
}
\value{
A \link{RelDataModel}
}
\description{
Remove a table from a \link{RelDataModel}
}
