% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{removeTable.RelDataModel}
\alias{removeTable.RelDataModel}
\title{Remove a table from a \link{RelDataModel}}
\usage{
\method{removeTable}{RelDataModel}(x, tableName)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{tableName}{the name of the table to remove}
}
\value{
A \link{RelDataModel}
}
\description{
Remove a table from a \link{RelDataModel}
}
