% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{rename_field}
\alias{rename_field}
\title{Rename an existing field in a \link{RelDataModel} table}
\usage{
rename_field(x, tableName, current, new)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{tableName}{the name of the table to modify (a single character)}

\item{current}{the current name of the field to modify (a single character)}

\item{new}{the new name of the field (a single character)}
}
\value{
A \link{RelDataModel}
}
\description{
Rename an existing field in a \link{RelDataModel} table
}
