% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{rename_table.RelDataModel}
\alias{rename_table.RelDataModel}
\title{Rename a table in a \link{RelDataModel}}
\usage{
\method{rename_table}{RelDataModel}(x, old, new)
}
\arguments{
\item{x}{a \link{RelDataModel} object}

\item{old}{a single character corresponding to the table name to change}

\item{new}{the new table name}
}
\value{
A \link{RelDataModel}
}
\description{
Rename a table in a \link{RelDataModel}
}
