% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{set_primary_key.RelDataModel}
\alias{set_primary_key.RelDataModel}
\title{Set the primary key a table in a \link{RelDataModel}}
\usage{
\method{set_primary_key}{RelDataModel}(x, tableName, fieldNames)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{tableName}{the name of the table to modify (a single character)}

\item{fieldNames}{the names of the fields to include in the primary key}
}
\value{
A \link{RelDataModel}
}
\description{
Set the primary key a table in a \link{RelDataModel}
}
