% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{set_unique_index}
\alias{set_unique_index}
\title{Set table index uniqueness in a \link{RelDataModel}}
\usage{
set_unique_index(x, tableName, fieldNames, unique)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{tableName}{the name of the table to modify (a single character)}

\item{fieldNames}{the names of the fields composing the index}

\item{unique}{a logical value}
}
\value{
A \link{RelDataModel}
}
\description{
Set table index uniqueness in a \link{RelDataModel}
}
