% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{[.RelDataModel}
\alias{[.RelDataModel}
\title{Extract from RelDataModel}
\usage{
\method{[}{RelDataModel}(x, i, rmForeignKeys = FALSE, ...)
}
\arguments{
\item{x}{the RelDataModel objct}

\item{i}{the index or the names of the elements to extract}

\item{rmForeignKeys}{if TRUE, remove foreign keys which are not
available after extraction. If FALSE (default) the function will throw an
error if any foreign keys does not exist in the extracted RelDataModel.}

\item{...}{additional arguments for the \code{\link{RelDataModel}} function.}
}
\description{
Extract from RelDataModel
}
