% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TYPEDEF.R
\name{typeRefConv}
\alias{typeRefConv}
\title{Convert a set of types from or to R supported types}
\usage{
typeRefConv(x, from = NULL, to = NULL)
}
\arguments{
\item{x}{a character vector of types to be converted. If from is not null,
x should be a set of valid types in the from reference. If to is not null,
x should be a set of supported R types (SUPPTYPES).}

\item{from}{a character vector of length one: the type reference
(\code{\link{listTypeRef}}) of x}

\item{to}{a character vector of length one: the targeted type reference
(\code{\link{listTypeRef}})}
}
\description{
Convert a set of types from or to R supported types
}
\details{
Only from XOR to should be set
}
