% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{update_field}
\alias{update_field}
\title{Update field information in a table of a \link{RelDataModel}}
\usage{
update_field(
  x,
  tableName,
  fieldName,
  type = NULL,
  nullable = NULL,
  unique = NULL,
  comment = NULL
)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{tableName}{the name of the table to modify (a single character)}

\item{fieldName}{the name of the field to modify (a single character)}

\item{type}{the type of the field (a single character)}

\item{nullable}{if the field is nullable (a single logical)}

\item{unique}{if the values are unique (a single logical)}

\item{comment}{a description (a single character)}
}
\value{
A \link{RelDataModel}
}
\description{
Update field information in a table of a \link{RelDataModel}
}
