% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{update_foreign_key}
\alias{update_foreign_key}
\title{Update a the cardinalities of a foreign key between two tables}
\usage{
update_foreign_key(
  x,
  fromTable,
  fromFields,
  toTable,
  toFields,
  fmin,
  fmax,
  tmin,
  tmax
)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{fromTable}{the name of the referencing table}

\item{fromFields}{the name of the referencing fields}

\item{toTable}{the name of the referenced table}

\item{toFields}{the names of the referenced fields}

\item{fmin}{from minimum cardinality}

\item{fmax}{from maximum cardinality}

\item{tmin}{to minimum cardinality}

\item{tmax}{to maximum cardinality}
}
\value{
A \link{RelDataModel}
}
\description{
Update a the cardinalities of a foreign key between two tables
}
