% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{update_table_display}
\alias{update_table_display}
\title{Update the display of a table of a \link{RelDataModel}}
\usage{
update_table_display(
  x,
  tableName,
  px = NULL,
  py = NULL,
  color = NULL,
  comment = NULL
)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{tableName}{the name of the table to modify (a single character)}

\item{px}{the position of the table: x value}

\item{py}{the position of the table: y value}

\item{color}{the color of the table}

\item{comment}{a table description/comment}
}
\value{
A \link{RelDataModel}
}
\description{
Update the display of a table of a \link{RelDataModel}
}
