% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{update_table_display.RelDataModel}
\alias{update_table_display.RelDataModel}
\title{Update the display of a table of a \link{RelDataModel}}
\usage{
\method{update_table_display}{RelDataModel}(x, tableName, px = NA, py = NA, color = NA, comment = NA)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{tableName}{the name of the table to modify (a single character)}

\item{px}{the position of the table: x value}

\item{py}{the position of the table: y value}

\item{color}{the color of the table}

\item{comment}{a table description/comment}
}
\value{
A \link{RelDataModel}
}
\description{
Update the display of a table of a \link{RelDataModel}
}
