% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_confrontation_report.R
\name{view_confrontation_report}
\alias{view_confrontation_report}
\title{View confrontation report in rstudio viewer}
\usage{
view_confrontation_report(cr, ...)
}
\arguments{
\item{cr}{the confrontation report from \link{confront_data.RelDataModel}}

\item{...}{additional params for the \link{format_confrontation_report} function}

\item{title}{report title}
}
\description{
View confrontation report in rstudio viewer
}
