% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_data_model.R
\name{write_json_data_model}
\alias{write_json_data_model}
\title{Write a data model in a JSON file}
\usage{
write_json_data_model(x, path)
}
\arguments{
\item{x}{the model to be written}

\item{path}{file on disk}
}
\description{
Write a data model in a JSON file
}
