% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDataModel.R
\name{check_foreign_keys}
\alias{check_foreign_keys}
\title{Check the availability of foreign keys}
\usage{
check_foreign_keys(x)
}
\arguments{
\item{x}{a \link{RelDataModel} object}
}
\value{
Nothing. The function throws an error if there is an issue with
foreign keys.
}
\description{
Check the availability of foreign keys
}
