% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelTableModel.R
\name{correct_constraints}
\alias{correct_constraints}
\title{Correct the constraints of a table to make them consistent}
\usage{
correct_constraints(x)
}
\arguments{
\item{x}{a \link{RelTableModel} object}
}
\description{
Correct the constraints of a table to make them consistent
}
