% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelTableModel.R
\name{format.RelTableModel}
\alias{format.RelTableModel}
\title{Format a \link{RelTableModel} object for printing}
\usage{
\method{format}{RelTableModel}(x, ...)
}
\arguments{
\item{x}{a \link{RelTableModel} object}

\item{...}{for generics compatibility (not used)}
}
\value{
A single character
}
\description{
Format a \link{RelTableModel} object for printing
}
