% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_confrontation_report.R
\name{format_confrontation_report}
\alias{format_confrontation_report}
\title{Format confrontation report for printing in console}
\usage{
format_confrontation_report(cr, title = "Model")
}
\arguments{
\item{cr}{the confrontation report from \link{confront_data}}

\item{title}{a character with a single value corresponding to the report
title (e.g. database/model name)}
}
\description{
Format confrontation report for printing in console
}
