% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_confrontation_report.R
\name{format_confrontation_report_md}
\alias{format_confrontation_report_md}
\title{Format confrontation report in markdown format}
\usage{
format_confrontation_report_md(
  cr,
  title = "Model",
  bgSuccess = "green",
  txSuccess = "black",
  bgFailure = "red",
  txFailure = "white",
  bgMessage = "#FFBB33",
  txMessage = "white"
)
}
\arguments{
\item{cr}{the confrontation report from \link{confront_data.RelDataModel}}

\item{title}{a character with a single value corresponding to the report}
}
\description{
Format confrontation report in markdown format
}
