% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_relational_data.R
\name{model_relational_data}
\alias{model_relational_data}
\title{Relational data modeler GUI}
\usage{
model_relational_data(modelInput = RelDataModel(list()), fromR = interactive())
}
\arguments{
\item{modelInput}{the \link{RelDataModel} to start from}

\item{fromR}{a logical indicating if the application is launched from R}
}
\description{
Relational data modeler GUI
}
