% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_relational_data.R
\name{model_relational_data}
\alias{model_relational_data}
\title{Relational data modeler GUI}
\usage{
model_relational_data(
  modelInput = RelDataModel(list()),
  fromR = interactive(),
  defaultColor = "#D9D9D9",
  availableColors = c("#9BC8FE", "#F67FC4", "#C6BDF1", "#DFFB86", "#F8DEC3", "#8FE6E0",
    "#FEFE8F", "#FAC6DC", "#A9ECC9"),
  example = system.file("examples/HPO-model.sql", package = packageName())
)
}
\arguments{
\item{modelInput}{the \link{RelDataModel} to start from}

\item{fromR}{a logical indicating if the application is launched from R}

\item{defaultColor}{a single color indicating the default table color}

\item{availableColors}{a character of possible colors for tables}

\item{example}{a file path to an sql or json model}
}
\description{
Relational data modeler GUI
}
