% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RelDataModel.R
\name{modelToVn}
\alias{modelToVn}
\title{VisNetwork representation of a \link{RelDataModel} object}
\usage{
modelToVn(
  model,
  color = "lightgrey",
  border = "black",
  highlightBorder = "orange"
)
}
\arguments{
\item{model}{a \link{RelDataModel}}

\item{color}{default table background color}

\item{border}{border color (single character)}

\item{highlightBorder}{color of highlighted borders

Internal function}
}
\description{
VisNetwork representation of a \link{RelDataModel} object
}
