% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RelDataModel.R
\name{plot.RelDataModel}
\alias{plot.RelDataModel}
\title{Plot a \link{RelDataModel} object}
\usage{
\method{plot}{RelDataModel}(x, reproducible_layout = TRUE, ...)
}
\description{
This function draw a visNetwork of the model.
}
