% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RelDataModel.R
\name{plot.RelDataModel}
\alias{plot.RelDataModel}
\title{Plot a \link{RelDataModel} object}
\usage{
\method{plot}{RelDataModel}(x, ...)
}
\arguments{
\item{x}{a \link{RelDataModel}}

\item{...}{additional parameters:
\itemize{
\item \strong{color} default table background color
\item \strong{border} border color (single character)
\item \strong{highlightBorder} color of highlighted borders
}}
}
\description{
This function draw a visNetwork of the \link{RelDataModel}.
}
\examples{
## Read the model ----
hpo_model <- read_json_data_model(
   system.file("examples/HPO-model.json", package="ReDaMoR")
)
## Plot the model ----
plot(hpo_model)
}
